refactor(bootstrap): converge assembly into wiring/wire; relocate network_policy
Phase 4.5 (final) of the uc-bootstrap redesign. Rename the remaining
assembly.rs to wiring/wire.rs and relocate network_policy.rs into wiring/.

wire.rs keeps the InfraLayer construction co-located with wire_dependencies:
the orchestrator consumes InfraLayer (and the SecureStoragePrelude /
SearchAssembly / CipherDecorators / BlobProcessingAssembly intermediates)
field-by-field, so they are one cohesive wiring unit. Splitting infra into a
separate module would only expose ~50 internal wiring fields as pub(crate)
with no encapsulation gain — a convergence the redesign §9.4 pre-authorized.

is_v2_blob moves next to its sole caller in layer/platform.rs (now private).
The composition root no longer has a flat assembly god module; the contract
wire_dependencies re-export sources from wiring::wire. No behavior change.
