+++ crates/uc-bootstrap/src/entrypoint/cli.rs
@@ -9,9 +9,9 @@
 
 use uc_core::config::AppConfig;
 
-use crate::assembly::wire_dependencies;
 use crate::layer::paths::get_storage_paths;
 use crate::wiring::deps::BackgroundRuntimeDeps;
+use crate::wiring::wire::wire_dependencies;
 
 /// Shared core wiring for the CLI composition-root entry.
 /// Initializes tracing, resolves config, wires dependencies, and registers the
