+++ crates/uc-bootstrap/src/entrypoint/daemon.rs
@@ -99,7 +99,7 @@ pub async fn build_daemon_lifecycle(
     // 【checker BLOCKER 4 — 单一取反点铁律】
     // `disable_relays` 的值**只能**通过 `relay_policy_to_iroh_config` 取得,
     // **不**在此处内联写 `let disable_relays = !allow_relay_fallback;`。
-    let mut iroh_config = crate::network_policy::relay_policy_to_iroh_config(
+    let mut iroh_config = crate::wiring::network_policy::relay_policy_to_iroh_config(
         allow_relay_fallback,
         allow_overlay_network_addrs,
         custom_relay_urls,
@@ -108,8 +108,8 @@ pub async fn build_daemon_lifecycle(
     );
     // #900：从 env 读取直连可达性（固定 UDP 端口 + 广播公网地址）并写入。
     // 必须在 `build_sync_engine_assembly`（首次 endpoint 快照/配对交换）之前。
-    crate::network_policy::apply_iroh_direct_reachability_from_env(&mut iroh_config);
-    crate::network_policy::apply_congestion_controller_from_env(&mut iroh_config);
+    crate::wiring::network_policy::apply_iroh_direct_reachability_from_env(&mut iroh_config);
+    crate::wiring::network_policy::apply_congestion_controller_from_env(&mut iroh_config);
 
     tracing::info!(
         target: "settings.network",
