+++ crates/uc-bootstrap/src/wiring/mod.rs
@@ -2,3 +2,5 @@
 //! `wire_dependencies` orchestrator that fills them.
 
 pub mod deps;
+pub(crate) mod network_policy;
+pub mod wire;
