+++ crates/uc-bootstrap/src/wiring/network_policy.rs
